refactor(demo): move TextInput examples to Input
The component is called Input so it should be displayed under a section that has that name.
